Move QoS models into a dedicated `models` package
